+++ modules/src/application/mod.rs
@@ -1,3 +1,2 @@
-
 pub mod error;
 pub mod msgs;
